Compile an algorithm as a shared library and python extension (#1)
Include a sample algorithm to show how it integrates into the build
file. The build file still needs to be modified to make it easy to
add multiple algorithms.

The python bindings currently require changing some files to reflect
the name of algorithms. We could generate these files automatically.
Importing the python package makes the algorithm available in
`essentia.standard`. Ideally it should be part of the other package,
but we need to rewrite some parts of the essentia bindings to allow
the use of a separate factory